+++ sources/fr.fmteam/res/source.json
@@ -1,0 +1,10 @@
+{
+  "info": {
+    "id": "fr.fmteam",
+    "name": "FMTeam",
+    "version": 1,
+    "url": "https://fmteam.fr",
+    "contentRating": 0,
+    "languages": ["fr"]
+  }
+}
