--- sources/fr.fmteamfr/res/source.json
@@ -1,10 +1,0 @@
-{
-	"info": {
-		"id": "fr.fmteamfr",
-		"name": "FMTeam.fr",
-		"version": 1,
-		"url": "https://fmteam.fr/",
-		"contentRating": 0,
-		"languages": ["fr"]
-	}
-}
